continue with structure refactoring and improving it and make example work by signing it.
