+++ packages/target-tauri/crates/user-notify/Cargo.toml
@@ -14,6 +14,7 @@ tokio = { version = "1.44.2", features = [
     "rt",
     "rt-multi-thread",
     "sync",
+    "signal",
 ] }
 url = "2.5.4"
