+++ packages/target-tauri/crates/user-notify/src/platform_impl/mac_os/mod.rs
@@ -1,0 +1,8 @@
+mod builder;
+mod delegate;
+mod handle;
+mod manager;
+
+pub use builder::NotificationBuilderMacOS;
+pub use handle::NotificationHandleMacOS;
+pub use manager::NotificationManagerMacOS;
