+++ packages/target-tauri/src-tauri/src/state/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod deltachat;
 pub(crate) mod html_email_instances;
 pub(crate) mod main_window_channels;
 pub(crate) mod menu_manager;
+pub(crate) mod notification;
 pub(crate) mod translations;
 #[cfg(desktop)]
 pub(crate) mod tray_manager;
